backend/requirements.txt
fastapi[standard]
python-dotenv
langchain_together
langchain
langchain-pinecone
supabase
pyjwt
passlib
bcrypt
